Validacoes questionario

+++ src/modules/questionario/questionario-service.ts
@@ -259,21 +259,34 @@ export function excluir(id: number) : Promise<ResultadoServico> {
           return dbResolve(new ResultadoServico('Questionário não encontrado', StatusServico.Erro));
         }
 
-        db.questionarioPerguntas.destroy({
+        return db.eventosQuestionarios.findAll({
           where: {
             idQuestionario: questionario.id
           },
           transaction: t
         })
-        .then(() => {
-          db.questionarios.destroy({
+        .then(resultado => {
+          
+          if (resultado.length > 0) {
+            return dbResolve(new ResultadoServico('Este questionário está vinculado á um evento. Remova-o do evento antes de excluí-lo.', StatusServico.Erro));
+          }
+
+          return db.questionarioPerguntas.destroy({
             where: {
-              id: questionario.id
+              idQuestionario: questionario.id
             },
             transaction: t
           })
-          .then(resp => {
-            dbResolve(new ResultadoServico(resp));
+          .then(() => {
+            return db.questionarios.destroy({
+              where: {
+                id: questionario.id
+              },
+              transaction: t
+            })
+            .then(resp => {
+              dbResolve(new ResultadoServico(resp));
+            });
           });
         });
       })
@@ -332,7 +345,7 @@ export function vincularPergunta(idQuestionario: number, idPergunta: number) : P
   });
 };
 
-export function desvincularPergunta(idQuestionario: number, idPergunta: number) : Promise<ResultadoServico> {
+export function validarDesvinculoPergunta(idQuestionario: number, idPergunta: number) : Promise<ResultadoServico> {
   return new Promise((resolve, reject) => {
 
     Promise.all([
@@ -356,19 +369,18 @@ export function desvincularPergunta(idQuestionario: number, idPergunta: number) 
         return resolve(new ResultadoServico('Pergunta não encontrada', StatusServico.Erro));
       }
 
-      db.questionarioPerguntas.destroy({
+      db.respostas.findAll({
         where: {
-          idPergunta: pergunta.id,
-          idQuestionario: questionario.id
+          idPergunta: pergunta.id
         }
-      })
-      .then((resp) => {
-        resolve(new ResultadoServico(resp));
-      })
-      .catch(err => {
-        reject(new ResultadoServico(err, StatusServico.Erro, TipoErro.Excecao));
-      });
+      }).then(resultado => {
 
+        if (resultado.length > 0) {
+          return resolve(new ResultadoServico('Esta pergunta já foi respondida neste ou em outro questionário. Não é mais possível removê-la.', StatusServico.Erro));
+        }
+        
+        resolve(new ResultadoServico(true));
+      })
     })
     .catch(err => {
       reject(new ResultadoServico(err, StatusServico.Erro, TipoErro.Excecao));
